Always an unsealed host to send API commands to
This allows us to run all steps even if one of the hosts is sealed for some
reason.

Under normal execution, this shouldn't happen since we always start by
unsealing all nodes in the cluster. However, if running only a subset of tasks
we need to make sure we don't accidentally run against a node which hasn't been
unsealed.