Support generating multiple unseal key shares per public key
This commit gets as far as enabling generating multiple keys per PGP key/user
but not so far that the role will actually work if you do this. So far we don't
ever submit more than one key and things get pretty confused about which key
that should be. This is work for a subsequent commit!

This is being done to work-around Vault's insistence that thresholds of >= 2
must be used when the number of key shares is > 1...