Do not run generate_root or decrypt_unseal_keys in roles
Previously this was done not for convenience (though it sort-of is) but so that
unseal keys were not fetched and decrypted and root tokens not generated unless
actually needed.

This might have felt a little surprising for users of roles which did this. (Or
for those which didn't!) It also led to excessive quantites of 'skipped'
messages being produced whilst running most roles.

In the end, it turned out there wasn't much need for this and that instead
playbooks should just perform those tasks explicitly at the beginning before
moving on. And thats what this commit does.