Make Vault server systemd setup part of bbcrd.vault.configure_server
This means that 'bbcrd.vault.install' can be used in cases where only the Vault
CLI is needed.

This is a sort-of breaking change, but nobody is likely to be running the
install role without the configure_server role and hoping for a working server.
As such I'm going to brush this under the rug as a non-breaking change...